Move save images to inline
